// lib/lambda/src/functions/adminLogin/dtos/adminLogin.dto.ts
export interface LoginRequest {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  password: string;
  status: number;
  role_id: number;
  role_name: string;
}

export interface LoginResponse {
  user: {
    id: number;
    username: string;
    status: number;
    role_id: number;
    role_name: string;
  };
  tokens: {
    accessToken: string;
    refreshToken: string;
    expiresIn: number;
    refreshExpiresIn: number;
    tokenType: string;
  };
}

export interface JWTPayload {
  userId: number;
  username: string;
  roleId: number;
  roleName: string;
} 